deep-research-agent/pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Multi-agent deep research system"
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
    "tiktoken>=0.8.0",
    "faiss-cpu>=1.9.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "duckdb>=1.1.0",
    "nbformat>=5.10.0",
    "jupyter-client>=8.6.0",
    "nbconvert>=7.16.0",
    "scikit-learn>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
research = "deep_research.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.0",
    "types-seaborn>=0.13.0",
    "ipykernel>=7.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
